// backend/src/common/enums/user.enum.ts
export enum UserRole {
  SUPER_ADMIN = "SUPER_ADMIN",
  ADMIN = "ADMIN",
  INTERNAL_TEACHER = "INTERNAL_TEACHER",
  EXTERNAL_TEACHER = "EXTERNAL_TEACHER",
  INTERNAL_STUDENT = "INTERNAL_STUDENT",
  EXTERNAL_STUDENT = "EXTERNAL_STUDENT",
}

export enum UserStatus {
  ACTIVE = 'active',
  INACTIVE = 'inactive',
  SUSPENDED = 'suspended',
  PENDING = 'pending',
}